fix: use isize for skip/limit in get_range to prevent overflow
Change skip and limit parameters from usize to isize in:
- lists::get_range
- RedisOps::try_from_index_list trait method
- Event::get_events_from_redis
- EventsQuery and EventsList in events route

This prevents integer overflow that occurred when usize::MAX was cast
to isize for Redis LRANGE operations.

+++ nexus-common/src/models/event/mod.rs
@@ -104,9 +104,9 @@ impl Event {
     }
 
     pub async fn get_events_from_redis(
-        cursor: Option<usize>,
-        limit: usize,
-    ) -> Result<(Vec<String>, usize), DynError> {
+        cursor: Option<isize>,
+        limit: isize,
+    ) -> Result<(Vec<String>, isize), DynError> {
         let start = cursor.unwrap_or(0);
         let result = Event::try_from_index_list(&["Events"], Some(start), Some(limit)).await;
 
@@ -118,7 +118,7 @@ impl Event {
             }
         };
 
-        let next_cursor = start + events.len();
+        let next_cursor = start + events.len() as isize;
 
         Ok((events, next_cursor))
     }
